rteval: pi_stress: Change PRIu64 to llu to match upstream sched_attr
- Change from PRIu64 to llu to match upstream glibc sched_attr.
This change is for when we are getting the definitions from upstream
glibc

- Change our definition in rt-sched.h to match upstream glibc
  This change is for when we are using our own definition of struct
sched_attr

+++ src/include/rt-sched.h
@@ -56,9 +56,9 @@ struct sched_attr {
 	uint32_t sched_priority;
 
 	/* SCHED_DEADLINE */
-	uint64_t sched_runtime;
-	uint64_t sched_deadline;
-	uint64_t sched_period;
+	unsigned long long sched_runtime;
+	unsigned long long sched_deadline;
+	unsigned long long sched_period;
 };
 
 int sched_setattr(pid_t pid,
